Extract code action resolve to a service
